EnergyModelsRenewableProducers.jl

EnergyModelsRenewableProducers — Module

Main module for EnergyModelsRenewableProducers.jl.

This module implements the following types (Nodes) with constraints:

NonDisRes is a subtype of Source and represents a non-dispatchable renewable producer, as wind, solar etc.

PumpedHydroStor is a subtype of Storage and represents a regulated pumped hydro storage.

HydroStor is a subtype of Storage and represents a regulated hydro storage, that is a standard hydro powerplant without pumps.

This Julia package implements two main nodes with corresponding JuMP constraints, extending the package EnergyModelsBase.jl with more detailed representation of renewable energy sources.

The first node, NonDisRES, models a non-dispatchable renewable energy source, like wind power, solar power, or run of river hydropower. These all use intermittent energy sources in the production of energy, so the maximum production capacity varies with the availability of the energy source at the time.

The other node implements a regulated hydropower storage plant, both with (PumpedHydroStor) and without pumps (HydroStor) for filling the reservoir with excess energy. The hydropower storage plant can also be extended as they are declared as subtypes of HydroStorage.

The new introduced node types are also documented in the public library.

Manual outline

Quick Start

Optimization variables

NonDisRES

HydroStorage

Constraint functions

Storage level constraints

Operational expenditure constraints

Examples

The package is installed with ] add

The code was downloaded with git clone

Library outline

Public interface

NonDisRES (non-dispatchable renewable energy source)

HydroStorage (regulated hydro storage with or without pump)

Internals

Types

Methods