Methods - Accessing fields

Index

EnergyModelsRenewableProducers.level_inflow

EnergyModelsRenewableProducers.level_init

EnergyModelsRenewableProducers.level_min

EnergyModelsRenewableProducers.opex_var_pump

EnergyModelsRenewableProducers.profile

NonDisRES

EnergyModelsRenewableProducers.profile — Function

profile(n::NonDisRES)
profile(n::NonDisRES, t)

Returns the profile of a node n of type NonDisRES either as TimeProfile or at operational period t.

HydroStorage

EnergyModelsRenewableProducers.level_init — Function

level_init(n::HydroStorage)
level_init(n::HydroStorage, t)

Returns the initial level of a node n of type HydroStorage either as TimeProfile or at operational period t.

EnergyModelsRenewableProducers.level_inflow — Function

level_inflow(n::HydroStorage)
level_inflow(n::HydroStorage, t)

Returns the inflow to a node n of type HydroStorage either as TimeProfile or at operational period t.

EnergyModelsRenewableProducers.level_min — Function

level_min(n::HydroStorage)
level_min(n::HydroStorage, t)

Returns the minimum level of a node n of type HydroStorage either as TimeProfile or at operational period t.

PumpedHydroStor

EnergyModelsRenewableProducers.opex_var_pump — Function

opex_var_pump(n::PumpedHydroStor)
opex_var_pump(n::PumpedHydroStor, t)

Returns the variable OPEX of a node n of type PumpedHydroStor related to pumping either as TimeProfile or at operational period t.