Types

Index

EnergyModelsRenewableProducers.AbstractBatteryLife

EnergyModelsRenewableProducers.AbstractPqCurve

EnergyModelsRenewableProducers.AbstractScheduleType

EnergyModelsRenewableProducers.HydroUnit

Nodal supertypes

EnergyModelsRenewableProducers.HydroUnit — Type

abstract type HydroUnit <: EMB.NetworkNode

A Hydropower unit node for either pumping or production, modelled as a NetworkNode node.

Parameter supertypes

EnergyModelsRenewableProducers.AbstractScheduleType — Type

abstract type AbstractScheduleType

Abstract supertype for the different constraint types.

EnergyModelsRenewableProducers.AbstractPqCurve — Type

abstract type AbstractPqCurve

AbstractPqCurve type used to represent the relationship between discharge of water and generation of electricity.

EnergyModelsRenewableProducers.AbstractBatteryLife — Type

AbstractBatteryLife

Abstract supertype for the modelling of the battery lifetime of an AbstractBattery. It allows to differentiate between different degradation approaches for the storage node.