Methods - Internal

Index

EnergyModelsRenewableProducers.build_hydro_reservoir_vol_constraints

EnergyModelsRenewableProducers.build_pq_constaints

EnergyModelsRenewableProducers.build_schedule_constraint

EnergyModelsRenewableProducers.has_penalty

EnergyModelsRenewableProducers.has_penalty_down

EnergyModelsRenewableProducers.has_penalty_up

EnergyModelsRenewableProducers.is_active

EnergyModelsRenewableProducers.is_constraint_data

EnergyModelsRenewableProducers.is_constraint_resource

Constraint functions

EnergyModelsRenewableProducers.build_hydro_reservoir_vol_constraints — Function

build_hydro_reservoir_vol_constraints(m, n::HydroReservoir, c::ScheduleConstraint, 𝒯)

Create minimum/maximum/schedule volume constraints for a HydroReservoir node. The ScheduleConstraint{T} can have types T <: AbstractScheduleType that defines the direction of the constraint.

Penalty variables are included unless the penalty value is not set or Inf.

EnergyModelsRenewableProducers.build_pq_constaints — Function

build_pq_constaints(m, n::HydroUnit, c::PqPoints, 𝒯::TimeStructure)

Function for creating the constraints on the variables flow_out, cap_use, and discharge_segments as declared in the PqPoints pq of a HydroUnit node.

EnergyModelsRenewableProducers.build_schedule_constraint — Function

build_schedule_constraint(m, n::Union{HydroGate, HydroUnit}, c::ScheduleConstraint, 𝒯::TimeStructure, p::ResourceCarrier)

Create minimum/maximum/schedule discharge constraints for the generic Node type. The ScheduleConstraint{T} can have types T <: AbstractScheduleType that defines the direction of the constraint. Penalty variables are included unless penalty value is not set or Inf`.

Identification functions

EnergyModelsRenewableProducers.is_constraint_data — Function

is_constraint_data(data::Data)
is_constraint_data(data::ScheduleConstraint)

Returns true if Data input is of type ScheduleConstraint, otherwise false.

EnergyModelsRenewableProducers.is_constraint_resource — Function

is_constraint_resource(data::ScheduleConstraint, resource::Resource)

Returns true if Data is of type ScheduleConstraint and ScheduleConstraint resource type is resource.

EnergyModelsRenewableProducers.is_active — Function

is_active(data::ScheduleConstraint, t)

Returns true if given constraint data is active at operational period t.

EnergyModelsRenewableProducers.has_penalty — Function

has_penalty(data::ScheduleConstraint, t)

Returns true if a constraint needs a penalty variable at operational period t.

EnergyModelsRenewableProducers.has_penalty_up — Function

has_penalty_up(data::ScheduleConstraint)
has_penalty_up(data::ScheduleConstraint, t)
has_penalty_up(data::ScheduleConstraint, t, p::Resource)

Returns true if a constraint data is of a type that may require a penalty up variable, which is true for MinSchedule and EqualSchedule.

When the operational period t is provided in addition, it is furthermore necessary that the penalty is finite.

When the operational period t and the resource p is provided in addition, it is furthermore necessary that the penalty is finite and that p corresponds to the ScheduleConstraint resource.

EnergyModelsRenewableProducers.has_penalty_down — Function

has_penalty_down(data::ScheduleConstraint)
has_penalty_down(data::ScheduleConstraint, t)
has_penalty_down(data::ScheduleConstraint, t, resource::Resource)

Returns true if a constraint data is of a type that may require a penalty up down, which is true for MaxSchedule and EqualSchedule.

When the operational period t is provided in addition, it is furthermore necessary that the penalty is finite.

When the operational period t and the resource p is provided in addition, it is furthermore necessary that the penalty is finite and that p corresponds to the ScheduleConstraint resource.