Methods - Accessing fields

Index

EnergyModelsRenewableProducers.battery_life

EnergyModelsRenewableProducers.constraint_data

EnergyModelsRenewableProducers.cycles

EnergyModelsRenewableProducers.degradation

EnergyModelsRenewableProducers.discharge_level

EnergyModelsRenewableProducers.discharge_segments

EnergyModelsRenewableProducers.electricity_resource

EnergyModelsRenewableProducers.level_inflow

EnergyModelsRenewableProducers.level_init

EnergyModelsRenewableProducers.level_min

EnergyModelsRenewableProducers.max_normalized_flow

EnergyModelsRenewableProducers.max_normalized_power

EnergyModelsRenewableProducers.opex_var_pump

EnergyModelsRenewableProducers.penalty

EnergyModelsRenewableProducers.power_level

EnergyModelsRenewableProducers.pq_curve

EnergyModelsRenewableProducers.profile

EnergyModelsRenewableProducers.reserve_down

EnergyModelsRenewableProducers.reserve_up

EnergyModelsRenewableProducers.resource

EnergyModelsRenewableProducers.stack_cost

EnergyModelsRenewableProducers.value

EnergyModelsRenewableProducers.vol_inflow

EnergyModelsRenewableProducers.water_resource

Node

EnergyModelsRenewableProducers.constraint_data — Function

constraint_data(n::EMB.Node)

Returns vector of Data that are of type ScheduleConstraint.

NonDisRES

EnergyModelsRenewableProducers.profile — Function

profile(n::AbstractNonDisRES)
profile(n::AbstractNonDisRES, t)

Returns the profile of a node n of type AbstractNonDisRES either as TimeProfile or in operational period t.

HydroStorage

EnergyModelsRenewableProducers.level_init — Function

level_init(n::HydroStorage)
level_init(n::HydroStorage, t)

Returns the initial level of a node n of type HydroStorage either as TimeProfile or in operational period t.

EnergyModelsRenewableProducers.level_inflow — Function

level_inflow(n::HydroStorage)
level_inflow(n::HydroStorage, t)

Returns the inflow to a node n of type HydroStorage either as TimeProfile or in operational period t.

EnergyModelsRenewableProducers.level_min — Function

level_min(n::HydroStorage)
level_min(n::HydroStorage, t)

Returns the minimum level of a node n of type HydroStorage either as TimeProfile or in operational period t.

PumpedHydroStor

EnergyModelsRenewableProducers.opex_var_pump — Function

opex_var_pump(n::PumpedHydroStor)
opex_var_pump(n::PumpedHydroStor, t)

Returns the variable OPEX of a node n of type PumpedHydroStor related to pumping either as TimeProfile or in operational period t.

HydroReservoir

EnergyModelsRenewableProducers.max_normalized_power — Function

max_normalized_power(n::HydroUnit)

Returns the maximum power as a fraction of the capacity of HydroUnit n based on the pq_curve input.

EnergyModelsRenewableProducers.max_normalized_flow — Function

max_normalized_flow(n::HydroUnit)

Returns the maximum flow as a fraction of the capacity of HydroUnit n based on the pq_curve input.

EnergyModelsRenewableProducers.vol_inflow — Function

vol_inflow(n::HydroReservoir)
vol_inflow(n::HydroReservoir, t)

Returns the inflow to a HydroReservoir n either as TimeProfile or in operational period t.

HydroUnit

EnergyModelsRenewableProducers.pq_curve — Function

pq_curve(n::HydroUnit)

Returns the resources in the PQ-curve of a HydroUnit n.

EnergyModelsRenewableProducers.water_resource — Function

water_resource(n::HydroUnit)

Returns the resource of the water_resource field of a HydroUnit n.

EnergyModelsRenewableProducers.electricity_resource — Function

electricity_resource(n::HydroUnit)

Returns the resource of the electricity_resource field of a HydroUnit n.

Battery fields

EnergyModelsRenewableProducers.battery_life — Function

battery_life(n::AbstractBattery)

Returns the AbstractBatteryLife type of AbstractBattery n.

EnergyModelsRenewableProducers.cycles — Function

cycles(n::AbstractBattery)

Returns the maximum number of cycles of AbstractBattery n through calling its subfunction. If the battery_life is an AbstractBatteryLife, it will return nothing.

EnergyModelsRenewableProducers.degradation — Function

degradation(n::AbstractBattery)

Returns the degradation of the battery storage capacity at the end of its lifetime through calling its subfunction. If the battery_life is an AbstractBatteryLife, it will return nothing.

EnergyModelsRenewableProducers.stack_cost — Function

stack_cost(n::AbstractBattery)
stack_cost(n::AbstractBattery, t_inv::TS.AbstractStrategicPeriod)

Returns the relative stack cost of the battery storage capacity for replacing the existing battery capacity as TimeProfile or in strategic period t_inv.

If the battery_life is an AbstractBatteryLife, it will return nothing.

EnergyModelsRenewableProducers.reserve_up — Function

reserve_up(n::ReserveBattery)

Returns the instances used as reserve resources for adding capacity to the energy system.

EnergyModelsRenewableProducers.reserve_down — Function

reserve_down(n::ReserveBattery)

Returns the instances used as reserve resources for removing capacity from the energy system.

Scheduling constraints

EnergyModelsRenewableProducers.resource — Function

resource(data::ScheduleConstraint)

Returns the Resource type of a ScheduleConstraint.

EnergyModelsRenewableProducers.value — Function

value(data::ScheduleConstraint, t)

Returns the value of a constraint data at operational period t.

EnergyModelsRenewableProducers.penalty — Function

penalty(data::ScheduleConstraint, t)

Returns the penalty value of constraint data at operational period t.

PqPoints

EnergyModelsRenewableProducers.discharge_segments — Function

discharge_segments(pq_curve::PqPoints)

Returns the range of segment indices for a PqPoints pq_curve.

EnergyModelsRenewableProducers.discharge_level — Function

discharge_level(pq::PqPoints)
discharge_level(pq::PqPoints, i)

Returns the discharge level of PqPoint pq as array or at index i.

EnergyModelsRenewableProducers.power_level — Function

power_level(pq::PqPoints)
power_level(pq::PqPoints, i)

Returns the power level of PqPoint pq as array or at index i.