Types

Index

EnergyModelsRenewableProducers.AbstractPqCurve

EnergyModelsRenewableProducers.AbstractScheduleType

EnergyModelsRenewableProducers.EqualSchedule

EnergyModelsRenewableProducers.HydroUnit

EnergyModelsRenewableProducers.MaxSchedule

EnergyModelsRenewableProducers.MinSchedule

EnergyModelsRenewableProducers.PqPoints

EnergyModelsRenewableProducers.ScheduleConstraint

Nodal supertypes

EnergyModelsRenewableProducers.HydroUnit — Type

abstract type HydroUnit <: EMB.NetworkNode

A Hydropower unit node for either pumping or production, modelled as a NetworkNode node.

Constraint types

EnergyModelsRenewableProducers.AbstractScheduleType — Type

abstract type AbstractScheduleType

Abstract supertype for the different constraint types.

EnergyModelsRenewableProducers.MinSchedule — Type

abstract type MinSchedule <: AbstractScheduleType

Abstract type used to define a ScheduleConstraint as a minimum constraint.

EnergyModelsRenewableProducers.MaxSchedule — Type

abstract type MaxSchedule <: AbstractScheduleType

Abstract type used to define a ScheduleConstraint as a maximum constraint.

EnergyModelsRenewableProducers.EqualSchedule — Type

abstract type EqualSchedule <: AbstractScheduleType

Abstract type used to define a ScheduleConstraint as a schedule constraint.

EnergyModelsRenewableProducers.ScheduleConstraint — Type

ScheduleConstraint{T} <: Data where {T<:AbstractScheduleType}

A constraint that can be added as Data. T <: AbstractScheduleType denotes the constraint type.

Fields

resource::{Union{<:Resource, Nothing}} is the resource type the constraint applies to if the node can have multiple resources as input/outputs.

value::TimeProfile is the constraint value, that is the limit that should not be violated.

flag::TimeProfile is a boolean value indicating if the constraint is active.

penalty::TimeProfile is the penalty for violating the constraint. If penalty is set to Inf it will be built as a hard constraint.

Power-flow curves

EnergyModelsRenewableProducers.AbstractPqCurve — Type

abstract type AbstractPqCurve

AbstractPqCurve type used to represent the relationship between discharge of water and generation of electricity.

EnergyModelsRenewableProducers.PqPoints — Type

struct PqPoints <: AbstractPqCurve
PqPoints(power_levels::Vector{Real}, discharge_levels::Vector{Real})
PqPoints(eq::Real)

The relationship between discharge/pumping of water and power generation/consumption represented by a set of discharge and power values (PQ-points).

Fields

power_levels::Vector{Real} is a vector of power values.

discharge_levels::Vector{Real} is a vector of discharge values.

The two vectors muct be of equal size and ordered so that the power and discharge values describes the conversion from energy (stored in the water) to electricity (power) for a HydroGenerator node or the conversion from electric energy to energy stored as water in the reservoirs for a HydroPump node.

The first value in each vector should be zero. Furthermore, the vectors should be relative to the installed capacity, so that either the power-vector or the discharge vector is in the range [0, 1].

If a single Real is provided as input, it constructs the two Arrays through the energy equivalent input. If this approach is used, the installed capacity of the node must refer to the power capacity of a HydroGenerator or HydroPump node.

Note

The described power-discharge relationship should be concave for a HydroGenerator node and convex for a HydroPump node.