Public interface

NonDisRES (non-dispatchable renewable energy source)

This type models both wind power, solar power, and run of river hydropower. These have in common that they generate power from an intermittent energy source, so they can have large variations in power output, based on the availability of the renewable source at the time. These power sources can be modelled using the same type NonDisRES. The new type is a subtype of EMB.Source. The new type only differs from its supertype through the field profile.

The field profile::TimeProfile is a dimensionless ratio (between 0 and 1) describing how much of the installed capacity is utilized at the current operational period. Therefore, when using NonDisRES to model some renewable source, the data provided to this field is what defines the intermittent characteristics of the source.

The NonDisRES node is modelled very similar to a regular EMB.Source} node. The only difference is how the intermittent nature of the non-dispatchable source is handled. The maximum power generation of the source in the operational period $t$ depends on the time-dependent Profile variable.

Note

If not needed, the production does not need to run at full capacity. The amount of energy not produced is computed using the non-negative optimization variable $\texttt{curtailment}$ (declared for NonDisRES nodes only).

The fields of the different types are listed below:

EnergyModelsRenewableProducers.NonDisRES — Type

NonDisRES <: EMB.Source

A non-dispatchable renewable energy source. It extends the existing RefSource node through including a profile that corresponds to thr production. The profile can have variations on the strategic level.

Fields

id is the name/identifyer of the node.

cap::TimeProfile is the installed capacity.

profile::TimeProfile is the power production in each operational period as a ratio of the installed capacity at that time.

opex_var::TimeProfile is the variable operating expense per energy unit produced.

opex_fixed::TimeProfile is the fixed operating expense.

output::Dict{Resource, Real} are the generated Resources, normally Power.

data::Vector{Data} is the additional data (e.g. for investments). The field data is conditional through usage of a constructor.

HydroStorage (regulated hydro storage with or without pump)

A hydropower plant is much more flexible than, e.g., a wind farm since the water can be stored for later use. Energy can be produced (almost) whenever it is needed. Some hydropower plants also have pumps installed. These are used to pump water into the reservoir when excess and cheap energy is available in the network. EnergyModelsRenewableProducers introduces hence two different types representing a regulated hydropower plant (HydroStor) and a pumped regulated hydropower plant (PumpedHydroStor) without a lower reservoir. Both types have a level and discharge capacity while a PumpedHydroStor also includes a charge capacity.

The variable level_init represents the initial energy available in the reservoir in the beginning of each investment period. The variable level_inflow describes the inflow into the reservoir (measured in energy units), while level_min is the allowed minimum storage level in the dam, given as a ratio of the installed storage capacity of the reservoir at every operational period. The required minimum level is enforced by NVE and varies over the year. The resources stored in the hydro storage is set as stor_res, similar to a regular EMB.RefStorage.

The five last parameters are used in the same way as in EMB.Storage. In the implementation of PumpedHydroStor, the values set in input represents a loss of energy when using the pumps. A value of 1 means no energy loss, while a value of 0 represents 100% energy loss of that inflow variable.

The fields of the different types are listed below:

EnergyModelsRenewableProducers.HydroStorage — Type

An abstract type for hydro storage nodes, with or without pumping.

EnergyModelsRenewableProducers.HydroStor — Type

HydroStor{T} <: HydroStorage{T}

A regulated hydropower storage, modelled as a Storage node. A regulated hydro storage node requires a capacity for the discharge and does not have a required inflow from the model, except for water inflow from outside the model, although it requires a field input.

Fields

id is the name/identifyer of the node.

level::EMB.UnionCapacity are the level parameters of the HydroStor node. Depending on the chosen type, the charge parameters can include variable OPEX and/or fixed OPEX.

discharge::EMB.UnionCapacity are the discharging parameters of the HydroStor node. Depending on the chosen type, the discharge parameters can include variable OPEX, fixed OPEX, and/or a capacity.

level_init::TimeProfile is the initial stored energy in the dam.

level_inflow::TimeProfile is the inflow of power per operational period.

level_min::TimeProfile is the minimum fraction of the reservoir capacity that has to remain in the HydroStorage node.

stor_res::ResourceCarrier is the stored Resource.

input::Dict{Resource, Real} are the input Resources. In the case of a HydroStor, this field can be left out.

output::Dict{Resource, Real} can only contain one entry, the stored resource.

data::Vector{Data} additional data (e.g. for investments). The field data is conditional through usage of a constructor.

EnergyModelsRenewableProducers.PumpedHydroStor — Type

PumpedHydroStor{T} <: HydroStorage{T}

A pumped hydropower storage, modelled as a Storage node. A pumped hydro storage node allows for storing energy through pumping water into the reservoir. The current implementation is a simplified node in which no lower reservoir is required. Instead, it is assumed that the reservoir has an infinite size.

A pumped hydro storage node requires a capacity for both charge and discharge to account for the potential to store energy in the form of potential energy.

Fields

id is the name/identifyer of the node.

charge::EMB.UnionCapacity are the charging parameters of the PumpedHydroStor node. Depending on the chosen type, the charge parameters can include variable OPEX, fixed OPEX, and/or a capacity.

level::EMB.UnionCapacity are the level parameters of the HydroStor node. Depending on the chosen type, the charge parameters can include variable OPEX and/or fixed OPEX.

discharge::EMB.UnionCapacity are the discharging parameters of the HydroStor node. Depending on the chosen type, the discharge parameters can include variable OPEX, fixed OPEX, and/or a capacity.

level_init::TimeProfile is the initial stored energy in the dam.

level_inflow::TimeProfile is the inflow of power per operational period.

level_min::TimeProfile is the minimum fraction of the reservoir capacity that has to remain in the HydroStorage node.

stor_res::ResourceCarrier is the stored Resource.

input::Dict{Resource, Real} are the input Resources.

output::Dict{Resource, Real} can only contain one entry, the stored resource.

data::Vector{Data} additional data (e.g. for investments). The field data is conditional through usage of a constructor.

In recent version increases, we changed the individual fields of the HydroStorage nodes as well as their types. Hence, we still incorporate legacy constructors that can be utilized when having a model in previous versions. However, we removed one legacy constructor as it is no longer required. Calling the constructor will provide you now with an error.

This legacy constructor is:

EnergyModelsRenewableProducers.RegHydroStor — Function

RegHydroStor(
    id::Any,
    rate_cap::TimeProfile,
    stor_cap::TimeProfile,
    has_pump::Bool,
    level_init::TimeProfile,
    level_inflow::TimeProfile,
    level_min::TimeProfile,
    opex_var::TimeProfile,
    opex_fixed::TimeProfile,
    stor_res::ResourceCarrier,
    input,
    output,
    Data,
)

Original Legacy constructor for a regulated hydropower storage, with or without pumping capabilities. This version is discontinued starting with Version 0.6.0. resulting in an error It is replaced with the two new types HydroStor and PumpedHydroStor to utilize the concept of multiple dispatch instead of logic.

See the documentation for further information regarding how you can translate your existing model to the new model.

Fields

id is the name/identifyer of the node.

rate_cap::TimeProfile is the installed installed rate capacity.

stor_cap::TimeProfile is the installed storage capacity in the dam.

has_pump::Bool states wheter the stored resource can flow in.

level_init::TimeProfile is the initial stored energy in the dam.

level_inflow::TimeProfile is the inflow of power per operational period.

level_min::TimeProfile is the minimum fraction of the reservoir capacity that has to remain in the HydroStorage node.

opex_var::TimeProfile are the variable operational expenses per GWh produced.

opex_fixed::TimeProfile are the fixed operational costs of the storage caacity.

stor_res::ResourceCarrier is the stored Resource.

input::Dict{Resource, Real} are the stored and used resources. The values in the Dict are ratios describing the energy loss when using the pumps.

output::Dict{Resource, Real} can only contain one entry, the stored resource.

data::Array{Data} additional data (e.g. for investments). This value is conditional through the application of a constructor.

See the section on how to update models for further information regarding how you can translate your existing model to the new model.